+++ ErsatzTV.Core/Interfaces/Jellyfin/IJellyfinMovieLibraryScanner.cs
@@ -1,12 +1,12 @@
 ﻿using ErsatzTV.Core.Domain;
+using ErsatzTV.Core.Jellyfin;
 
 namespace ErsatzTV.Core.Interfaces.Jellyfin;
 
 public interface IJellyfinMovieLibraryScanner
 {
     Task<Either<BaseError, Unit>> ScanLibrary(
-        string address,
-        string apiKey,
+        JellyfinConnectionParameters connectionParameters,
         JellyfinLibrary library,
         bool deepScan,
         CancellationToken cancellationToken);