+++ ErsatzTV.Core/Jellyfin/JellyfinConnectionParameters.cs
@@ -3,4 +3,7 @@
 namespace ErsatzTV.Core.Jellyfin;
 
 public record JellyfinConnectionParameters(string Address, string ApiKey, int MediaSourceId)
-    : MediaServerConnectionParameters;
+    : MediaServerConnectionParameters
+{
+    public string AuthorizationHeader => $"MediaBrowser Token={ApiKey}";
+}